Updating Robot Classes (#622)
* isort fixes

* update robot APIs, verify all tests pass

* fix base indexing for two wheeled robots, verify differential drive demo script works

* fix misc bugs cf. eric comments

* update JR2 wheel / axle lengths, remove outdated keybinding API in TwoWheelRobot

* fixing black isort issues

* fix black

+++ igibson/controllers/null_gripper_controller.py
@@ -49,7 +49,7 @@ class NullGripperController(ManipulationController):
         super().__init__(
             control_freq=control_freq,
             control_limits=control_limits,
-            joint_idx=np.array([]),  # no joints controlled
+            joint_idx=np.array([], dtype=np.int),  # no joints controlled
             command_input_limits=command_input_limits,
             command_output_limits=command_output_limits,
         )
